+++ src/main/java/kitchenpos/menugroup/respository/MenuGroupRepository.java
@@ -1,6 +1,6 @@
-package kitchenpos.menu.repository;
+package kitchenpos.menugroup.respository;
 
-import kitchenpos.menu.domain.MenuGroup;
+import kitchenpos.menugroup.domain.MenuGroup;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface MenuGroupRepository extends JpaRepository<MenuGroup, Long> {
