+++ src/test/java/kitchenpos/order/application/OrderServiceTest.java
@@ -13,8 +13,8 @@ import java.util.List;
 import java.util.NoSuchElementException;
 import kitchenpos.ServiceTest;
 import kitchenpos.menu.domain.Menu;
-import kitchenpos.menu.domain.MenuGroup;
 import kitchenpos.menu.domain.MenuProduct;
+import kitchenpos.menugroup.domain.MenuGroup;
 import kitchenpos.order.domain.Order;
 import kitchenpos.order.domain.OrderLineItem;
 import kitchenpos.order.domain.OrderStatus;
